+++ src/main/java/kr/co/ohgoodfood/service/users/UserServiceImpl.java
@@ -47,173 +47,6 @@ public class UserServiceImpl implements UsersService{
     private final UserMapper userMapper;
     private final AwsS3Config awsS3Config;
 
-    /**
-     * LocalDate.now()로 오늘픽업, 내일픽업, 매진, 마감 상태를 판별합니다.
-     *
-     * @param mainStore          : 판별이 필요한 데이터가 담긴 객체
-     * @return                   : PickupStatus ENUM 객체
-     */
-    @Override
-    public PickupStatus getPickupDateStatus(MainStore mainStore) {
-        LocalDate today = LocalDate.now();
-
-        // [마감] - store_status = N
-        if("N".equals(mainStore.getStore_status())){
-            return PickupStatus.CLOSED;
-        }else{
-            if(mainStore.getPickup_start() == null){
-                return PickupStatus.CLOSED;
-            }
-            LocalDate pickupDate = mainStore.getPickup_start().toLocalDateTime().toLocalDate();
-            // [매진] - amount = 0
-            if(mainStore.getAmount() == 0){
-                return PickupStatus.SOLD_OUT;
-            }else{
-                // [오늘픽업] 현재 날짜와 같음
-                if (pickupDate.isEqual(today)) {
-                    return PickupStatus.TODAY;
-                }
-                // [내일픽업] 현재 날짜 + 1과 같음
-                // else if (pickupDate.isEqual(today.plusDays(1))) {
-                //     return PickupStatus.TOMORROW;
-                // }
-                else {
-                    return PickupStatus.TOMORROW;
-                }
-            }
-        }
-        // throw new IllegalStateException();
-    }
-
-    /**
-     * LocalDate.now()로 오늘픽업, 내일픽업만을 판별합니다.
-     * Orders 페이지에서는 마감,매진 값은 필요 없기 때문에, 이것만을 판별하는 로직을 따로 만듭니다.
-     *
-     * @param userOrder        : 판별이 필요한 데이터가 담긴 객체
-     * @return                 : PickupStatus ENUM 객체
-     */
-    @Override
-    public PickupStatus getOrderPickupDateStatus(UserOrderDTO userOrder) {
-        LocalDate today = LocalDate.now();
-        LocalDate pickupDate = userOrder.getPickup_start().toLocalDateTime().toLocalDate();
-
-        // [오늘픽업] 현재 날짜와 같음
-        if (pickupDate.isEqual(today)) {
-            return PickupStatus.TODAY;
-        }
-//        // [내일픽업] 현재 날짜 + 1과 같음
-//        if (pickupDate.isEqual(today.plusDays(1))) {
-//            return PickupStatus.TOMORROW;
-//        }
-        //불안정 하긴 하지만, 이틀 뒤가 pick_up start인 경우가 없기 때문에 나머진 다 내일 픽업
-        return PickupStatus.TOMORROW;
-    }
-
-    /**
-     * |(구분자) 구분은 확장성을 위해 프론트 단에 위임
-     * 서버에서는 리스트에 담아서 보내도록 한다.
-     *
-     * @param mainStore          : 판별이 필요한 데이터가 담긴 객체
-     * @return                   : 카테고리 이름이 담긴 List
-     */
-    @Override
-    public List<String> getCategoryList(MainStore mainStore) {
-        List<String> category_list = new ArrayList<>();
-
-        if(mainStore.getCategory_bakery().equals("Y")){
-            category_list.add("빵 & 디저트");
-        }
-
-        if(mainStore.getCategory_fruit().equals("Y")){
-            category_list.add("과일");
-        }
-
-        if(mainStore.getCategory_salad().equals("Y")){
-            category_list.add("샐러드");
-        }
-
-        if(mainStore.getCategory_others().equals("Y")){
-            category_list.add("그 외");
-        }
-
-        return category_list;
-    }
-
-    /**
-     * 사용자의 OrderList를 가져오는 method
-     *
-     * @param userOrderFilter    : 세션에 접속한 사용자 id와 필터링을 위한 객체가 담겨있다.
-     * @return                   : 조회한 UserOrderList
-     */
-    @Override
-    public List<UserOrderDTO> getUserOrderList(UserOrderFilter userOrderFilter){
-        List<UserOrderDTO> orderList = userMapper.selectOrderList(userOrderFilter);
-
-        // userOrder에 pickup_status와 block_cancel 상태를 저장.
-        for(UserOrderDTO userOrder : orderList){
-            userOrder.setPickup_status(getOrderPickupDateStatus(userOrder));
-            //주문에 해당하는 포인트를 적립
-            userOrder.setPoint(getOrderPoint(userOrder));
-            userOrder.setBlock_cancel(getOrderBlockCancel(userOrder.getOrder_status(), userOrder.getReservation_end()));
-        }
-        return orderList;
-    }
-    /**
-     * 사용자의 구매 금액별 포인트를 설정하기 위한 메서드
-     *
-     * @param userOrder          : orderList 화면에 뿌릴 객체
-     * @return                   : 구매 금액의 1%에 해당하는 point
-     */
-    public int getOrderPoint(UserOrderDTO userOrder){
-        return (int) (userOrder.getPaid_price() * 0.01);
-    }
-
-    /**
-     * pickup_status가 오늘픽업 혹은 내일 픽업인 경우에, (즉, confirmed 상태) 한 시간 전에 취소 block 상태를 만들기 위함입니다.
-     * reservation_end -1이 NOW일때를 계산합니다.
-     *
-     * @param order_status       : order_status가 reservation인 경우에만 진행한다.
-     * @param reservation_end    : 예약 마감 한시간 전을 계산하기 위한 reservation_end
-     * @return                   : block_cancel 값을 설정하기 위해 boolean return
-     */
-    public boolean getOrderBlockCancel(String order_status, Timestamp reservation_end){
-        if(order_status.equals("reservation")){
-            Timestamp now = new Timestamp(System.currentTimeMillis());
-
-            long oneHourInMillis = 60L * 60L * 1000L; //한시간 계산
-            Timestamp reservationEndOneHourBefore = new Timestamp(reservation_end.getTime() - oneHourInMillis);
-
-            //reservation_end -1h < now < reservation_end
-            if (now.after(reservationEndOneHourBefore) && now.before(reservation_end)) {
-                return true;  // 마감 1시간 전 이내이면, 취소 블록하고 막는다.
-            }
-            return false;
-        }
-        return false; //오늘 픽업, 내일 픽업이 아니라면 이 block 변수는 false
-    }
-
-    /**
-     * 사용자가 선택한 order를 취소 처리 합니다.
-     *
-     * @param userOrderRequest   : 사용자 주문 상태 변경 처리에 필요한 DTO
-     * @return                   : UPDATE 쿼리가 잘 실행 되었는지 보기 위해 row return
-     */
-    @Override
-    @Transactional
-    public boolean updateUserOrderCancel(UserOrderRequest userOrderRequest){
-        userOrderRequest.setOrder_status("cancel");
-        userOrderRequest.setCanceld_from("user");
-
-        int updateOrderCnt = userMapper.updateOrderStatus(userOrderRequest);
-        int updateAmountCnt = userMapper.restoreProductAmount(userOrderRequest);
-
-        // 하나라도 오류가 발생할 경우, 롤백을 위해 exception throws
-        if (updateOrderCnt != 1 || updateAmountCnt != 1) {
-            throw new IllegalStateException("주문 취소 처리 중 오류 발생");
-        }
-        return true;
-    }
-
     /** 유저 정보 한 건 조회 */
     @Override
     public UserMypage getUserInfo(String userId) {
