+++ src/main/java/kr/co/ohgoodfood/service/users/UsersService.java
@@ -16,27 +16,6 @@ import kr.co.ohgoodfood.dto.UserMypage;
  * - 유지 보수 및 확장 편의성을 위해 interface로 구성한다.
  */
 public interface UsersService {
-    //[판별 로직] 오늘 픽업, 내일 픽업, 마감 판별 연결 로직
-    PickupStatus getPickupDateStatus(MainStore mainStore);
-
-    //[판별 로직] UserOrder에서는 마감과 매진은 필요 없으므로, 오늘 픽업, 내일 픽업만 판별하는 로직
-    PickupStatus getOrderPickupDateStatus(UserOrderDTO userOrder);
-
-    //[판별 로직] 카테고리 List<String> 저장 로직
-    List<String> getCategoryList(MainStore mainStore);
-
-    //[Controller 로직] 주문내역 Controller 연결 로직
-    List<UserOrderDTO> getUserOrderList(UserOrderFilter userOrderFilter);
-
-    //[판별 로직] 구매 금액별 point를 계산하기 위한 메서드
-    int getOrderPoint(UserOrderDTO userOrder);
-
-    //[판별 로직] reservation_end 한 시간 전에 주문취소를 막아두기 위한 상태 판별 로직
-    boolean getOrderBlockCancel(String order_status, Timestamp reservation_end);
-
-    //[Controller 로직] 주문내역 취소 Controller 연결 로직
-    boolean updateUserOrderCancel(UserOrderRequest userOrderRequest);
-
     /* 사용자 기본 정보 한 건 조회*/
     UserMypage getUserInfo(String userId);
